+++ libgeo/rmxn.c
@@ -1,5 +1,5 @@
 /* See rmxn.h. */
-/* Last edited on 2024-11-22 02:04:52 by stolfi */
+/* Last edited on 2024-11-22 05:33:18 by stolfi */
 
 #include <math.h>
 #include <stdio.h>
@@ -72,7 +72,7 @@ void rmxn_map_row (uint32_t m, uint32_t n, double *x, double *A, double *r)
             double newSum = sum + tcorr;
             corr = (newSum - sum) - tcorr;
             sum = newSum;
-            t += n;
+            t += (int32_t)n;
           }
         r[j] = sum;
       }
@@ -100,7 +100,7 @@ void rmxn_mul (uint32_t m, uint32_t p, uint32_t n, double *A, double *B, double 
     for (int32_t i = 0; i < m; i++)
       { for (int32_t j = 0; j < n; j++)
           { double sum = 0.0, corr = 0.0;
-            uint32_t t = j;
+            int32_t t = j;
             for (int32_t k = 0; k < p; k++)
               { double term = A[r+k]*B[t];
                 /* Kahan's summation: */
@@ -108,11 +108,11 @@ void rmxn_mul (uint32_t m, uint32_t p, uint32_t n, double *A, double *B, double 
                 double newSum = sum + tcorr;
                 corr = (newSum - sum) - tcorr;
                 sum = newSum;
-                t+= n;
+                t += (int32_t)n;
               }
             M[v] = sum; v++;
           }
-        r += p;
+        r += (int32_t)p;
       }
   }
 
@@ -131,9 +131,9 @@ void rmxn_mul_tr (uint32_t m, uint32_t n, uint32_t p, double *A, double *B, doub
                 sum = newSum;
               }
             M[v] = sum; v++;
-            s += p;
+            s += (int32_t)p;
           }
-        r += p;
+        r += (int32_t)p;
       }
   }
 
@@ -150,7 +150,7 @@ void rmxn_tr_mul (uint32_t p, uint32_t m, uint32_t n, double *A, double *B, doub
                 double newSum = sum + tcorr;
                 corr = (newSum - sum) - tcorr;
                 sum = newSum;
-                r += m; s += n;
+                r += (int32_t)m; s += (int32_t)n;
               }
             M[v] = sum; v++;
           }
@@ -163,7 +163,7 @@ double rmxn_det (uint32_t n, double *A)
     double det = 1.0;
     for (int32_t t = 0; t < n2; t++) { C[t] = A[t]; }
     gsel_triangularize(n, n, C, FALSE, 0.0);
-    for (int32_t t = 0; t < n2; t += n+1) { det *= C[t]; }
+    for (int32_t t = 0; t < n2; t += (int32_t)n+1) { det *= C[t]; }
     free(C);
     return det;
   }
@@ -175,14 +175,14 @@ double rmxn_inv (uint32_t n, double *A, double *M)
     double *C = rmxn_alloc(n,nC);
     /* Copy {A} into the left half of {C}, fill the right half with the identity: */
     for (int32_t i = 0; i < n; i++) 
-      { double *Cij = &(C[nC*i]); 
-        double *Aij = &(A[nA*i]);
+      { double *Cij = &(C[(int32_t)nC*i]); 
+        double *Aij = &(A[(int32_t)nA*i]);
         for (int32_t j = 0; j < nA; j++) { (*Cij) = (*Aij); Cij++; Aij++; }
         for (int32_t j = 0; j < nM; j++) { (*Cij) = (j == i ? 1.0 : 0.0); Cij++; }
       }
     gsel_triangularize(n, nC, C, TRUE, 0.0);
     double det = 1.0;
-    for (int32_t i = 0; i < n; i++) { det *= C[nC*i + i]; }
+    for (int32_t i = 0; i < n; i++) { det *= C[(int32_t)nC*i + i]; }
     if (det == 0.0)
       { /* Singular matrix: */
         double *Mij = M;
@@ -195,8 +195,8 @@ double rmxn_inv (uint32_t n, double *A, double *M)
       { gsel_diagonalize(n, nC, C);
         gsel_normalize(n, nC, C);
         for (int32_t i = 0; i < n; i++) 
-          { double *Cij = &(C[nC*i + n]); 
-            double *Mij = &(M[nM*i]);
+          { double *Cij = &(C[(int32_t)nC*i + (int32_t)n]); 
+            double *Mij = &(M[(int32_t)nM*i]);
             for (int32_t j = 0; j < nM; j++) { (*Mij) = (*Cij); Mij++; Cij++; }
           }
       }
@@ -224,11 +224,11 @@ double rmxn_inv_full (uint32_t n, double *A, double *M)
         /* Store its value in {biga}, its indices in {prow[k],pcol[k]}: */
         prow[k] = k;
         pcol[k] = k;
-        double biga = M[k*n+k];
+        double biga = M[k*(int32_t)n+k];
         for (int32_t i = k; i < n; i++)
           { for (int32_t j = k; j < n; j++)
-              { if (fabs (M[i*n+j]) > fabs (biga))
-                  { biga = M[i*n+j]; prow[k] = i; pcol[k] = j; }
+              { if (fabs (M[i*(int32_t)n+j]) > fabs (biga))
+                  { biga = M[i*(int32_t)n+j]; prow[k] = i; pcol[k] = j; }
               }
           }
  
@@ -242,9 +242,9 @@ double rmxn_inv_full (uint32_t n, double *A, double *M)
         { int32_t i = prow[k];
           if (i > k)
             { for (int32_t j = 0; j < n; j++)
-                { double hold = -M[k*n+j];
-                  M[k*n+j] = M[i*n+j];
-                  M[i*n+j] = hold;
+                { double hold = -M[k*(int32_t)n+j];
+                  M[k*(int32_t)n+j] = M[i*(int32_t)n+j];
+                  M[i*(int32_t)n+j] = hold;
                 }
             }
         }
@@ -253,30 +253,30 @@ double rmxn_inv_full (uint32_t n, double *A, double *M)
         { int32_t j = pcol[k];
           if (j > k)
             { for (int32_t i = 0; i < n; i++)
-                { double hold = -M[i*n+k];
-                  M[i*n+k] = M[i*n+j];
-                  M[i*n+j] = hold;
+                { double hold = -M[i*(int32_t)n+k];
+                  M[i*(int32_t)n+k] = M[i*(int32_t)n+j];
+                  M[i*(int32_t)n+j] = hold;
                 }
             }
         }
  
         /* Negate column {k} and divide by minus the pivot {biga}. */
-        for (int32_t i = 0; i < n; i++) { if (i != k) { M[i*n+k] /= -biga; } }
+        for (int32_t i = 0; i < n; i++) { if (i != k) { M[i*(int32_t)n+k] /= -biga; } }
  
         /* Reduce the matrix: */
         for (int32_t i = 0; i < n; i++)
           { if (i != k)
-              { double hold = M[i*n+k];
+              { double hold = M[i*(int32_t)n+k];
                 for (int32_t j = 0; j < n; j++)
-                  { if (j != k) { M[i*n+j] += hold * M[k*n+j]; } }
+                  { if (j != k) { M[i*(int32_t)n+j] += hold * M[k*(int32_t)n+j]; } }
               }
           }
  
         /* Divide row {k} by pivot {biga}: */
-        for (int32_t j = 0; j < n; j++) { if (j != k) { M[k*n+j] /= biga; } }
+        for (int32_t j = 0; j < n; j++) { if (j != k) { M[k*(int32_t)n+j] /= biga; } }
  
         /* Set the diagonal element to its reciprocal: */
-        M[k*n+k] = 1.0 / biga;
+        M[k*(int32_t)n+k] = 1.0 / biga;
          
         /* Shrink: */
         k++;
@@ -292,9 +292,9 @@ double rmxn_inv_full (uint32_t n, double *A, double *M)
         { int32_t i = prow[k];
           if (i > k)
             { for (int32_t j = 0; j < n; j++) 
-                { double hold = M[j*n+k];
-                  M[j*n+k] = -M[j*n+i];
-                  M[j*n+i] = hold;
+                { double hold = M[j*(int32_t)n+k];
+                  M[j*(int32_t)n+k] = -M[j*(int32_t)n+i];
+                  M[j*(int32_t)n+i] = hold;
                 }
             }
         }
@@ -303,9 +303,9 @@ double rmxn_inv_full (uint32_t n, double *A, double *M)
         { int32_t j = pcol[k];
           if (j > k)
             { for (int32_t i = 0; i < n; i++)
-                { double hold = M[k*n+i];
-                  M[k*n+i] = -M[j*n+i];
-                  M[j*n+i] = hold;
+                { double hold = M[k*(int32_t)n+i];
+                  M[k*(int32_t)n+i] = -M[j*(int32_t)n+i];
+                  M[j*(int32_t)n+i] = hold;
                 }
             }
         }
@@ -395,7 +395,7 @@ double rmxn_max_abs_elem(uint32_t m, uint32_t n, double *A)
   { double emax = 0.0;
     for (int32_t i = 0; i < m; i++)
       { for (int32_t j = 0; j < n; j++)
-          { double Mij = fabs(A[i*n + j]);
+          { double Mij = fabs(A[i*(int32_t)n + j]);
             if (Mij > emax) { emax = Mij; }
           }
       }
@@ -422,32 +422,32 @@ void rmxn_LT_inv_map_col(uint32_t m, double *L, double *y, double *r)
   { for (int32_t i = 0; i < m; i++)
       { double sum = y[i], corr = 0.0;
         for (int32_t j = 0; j < i; j++)
-          { double term = -L[m*i + j]*r[j];
+          { double term = -L[(int32_t)m*i + j]*r[j];
             /* Kahan's summation: */
             double tcorr = term - corr;
             double newSum = sum + tcorr;
             corr = (newSum - sum) - tcorr;
             sum = newSum;
           }
-        r[i] = sum/L[m*i + i];
+        r[i] = sum/L[(int32_t)m*i + i];
       }
   }
 
 void rmxn_LT_pos_div(uint32_t m, uint32_t n, double *A, double *L, double *M)
   { for (int32_t k = 0; k < m; k++)
       { /* Note hack for decreasing loop with uint counter. */
-        for (int32_t jp1 = n; jp1 > 0; jp1--)
+        for (int32_t jp1 = (int32_t)n; jp1 > 0; jp1--)
           { int32_t j = jp1-1;
-            double sum = A[n*k + j], corr = 0.0;
-            for (int32_t i = (uint32_t)j+1; i < n; i++)
-              { double term = -M[n*k + i]*L[n*i + j];
+            double sum = A[(int32_t)n*k + j], corr = 0.0;
+            for (int32_t i = j+1; i < n; i++)
+              { double term = -M[(int32_t)n*k + i]*L[(int32_t)n*i + j];
                 /* Kahan's summation: */
                 double tcorr = term - corr;
                 double newSum = sum + tcorr;
                 corr = (newSum - sum) - tcorr;
                 sum = newSum;
               }
-            M[n*k + j] = sum/L[n*j + j];
+            M[(int32_t)n*k + j] = sum/L[(int32_t)n*j + j];
           }
       }
   }
@@ -455,16 +455,16 @@ void rmxn_LT_pos_div(uint32_t m, uint32_t n, double *A, double *L, double *M)
 void rmxn_LT_pre_div(uint32_t m, uint32_t n, double *L, double *A, double *M)
   { for (int32_t k = 0; k < n; k++)
       { for (int32_t i = 0; i < m; i++)
-          { double sum = A[n*i + k], corr = 0.0;
+          { double sum = A[(int32_t)n*i + k], corr = 0.0;
             for (int32_t j = 0; j < i; j++)
-              { double term = -L[m*i + j]*M[n*j + k];
+              { double term = -L[(int32_t)m*i + j]*M[(int32_t)n*j + k];
                 /* Kahan's summation: */
                 double tcorr = term - corr;
                 double newSum = sum + tcorr;
                 corr = (newSum - sum) - tcorr;
                 sum = newSum;
               }
-            M[n*i + k] = sum/L[m*i + i];
+            M[(int32_t)n*i + k] = sum/L[(int32_t)m*i + i];
           }
       }
   }
@@ -491,32 +491,32 @@ void rmxn_cholesky(uint32_t n, double *A, double *L)
         for (int32_t j = 0; j < i; j++)
           { double sum = 0.0, corr = 0.0;
             for (int32_t k = 0; k < j; k++)
-              { double term = L[n*i + k]*L[n*j + k];
+              { double term = L[(int32_t)n*i + k]*L[(int32_t)n*j + k];
                 /* Kahan's summation formula: */
                 double tcorr = term - corr;
                 double newSum = sum + tcorr;
                 corr = (newSum - sum) - tcorr;
                 sum = newSum;
               }
-            double Ljj = L[n*j + j];
+            double Ljj = L[(int32_t)n*j + j];
             affirm(Ljj != 0.0, "zero element in diagonal");
-            L[n*i + j] = (A[n*i + j] - sum)/Ljj;
+            L[(int32_t)n*i + j] = (A[(int32_t)n*i + j] - sum)/Ljj;
           }
         
         /* Compute {L[i,i] = sqrt(A[i,i] - SUM{ L[i,j]^2 : j = 0..i-1 })} */
         double sum = 0.0, corr = 0.0;
         for (int32_t j = 0; j < i; j++)
-          { double w = L[n*i+j], term = w*w; 
+          { double w = L[(int32_t)n*i+j], term = w*w; 
             /* Kahan's summation formula: */
             double tcorr = term - corr;
             double newSum = sum + tcorr;
             corr = (newSum - sum) - tcorr;
             sum = newSum;
           }
-        double Lii = A[n*i+i] - sum;
+        double Lii = A[(int32_t)n*i+i] - sum;
         affirm (Lii >= 0.0, "matrix is not positive definite?");
-        L[n*i + i] = sqrt(Lii);
-        for (int32_t j = i+1; j < n; j++) { L[n*i + j] = 0.0; }
+        L[(int32_t)n*i + i] = sqrt(Lii);
+        for (int32_t j = i+1; j < n; j++) { L[(int32_t)n*i + j] = 0.0; }
       }
   }
 
@@ -524,7 +524,7 @@ void rmxn_perturb_unif(uint32_t m, uint32_t n, double pabs, double prel, double 
   {
     for (int32_t i = 0; i < m; i++)
       { for (int32_t j = 0; j < n; j++) 
-          { double *Aij = &(A[i*n + j]);
+          { double *Aij = &(A[i*(int32_t)n + j]);
             double mag = pabs + prel*fabs(*Aij);
             double d = dabrandom(-mag, +mag);
             (*Aij) += d; 
@@ -607,7 +607,7 @@ void rmxn_gen_print3
           { uint32_t nk = (k == 1 ? n1 : (k == 2 ? n2 : n3));
             double *Ak = (k == 1 ? A1 : (k == 2 ? A2 : A3));
             if (Ak != NULL)
-              { double *Aki = (nk == 0 ? NULL : &(Ak[i*nk]));
+              { double *Aki = (nk == 0 ? NULL : &(Ak[i*(int32_t)nk]));
                 if (k > 1) { fputs(msep, f); }
                 fputs(ilp, f);
                 for (int32_t j = 0; j < nk; j++)
