+++ libgeo/rmxn_shift.c
@@ -1,5 +1,5 @@
 /* See rmxn_shift.h. */
-/* Last edited on 2024-11-20 14:07:32 by stolfi */
+/* Last edited on 2024-11-22 05:37:41 by stolfi */
 
 #include <stdio.h>
 #include <stdint.h>
@@ -19,14 +19,14 @@
 void rmxn_shift_rows(uint32_t m, uint32_t n, double A[], double v[], double M[])
   { for (int32_t i = 0; i < m; i++)
       { for (int32_t j = 0; j < n; j++) 
-          { uint32_t ij = i*n + j; M[ij] = A[ij] + v[j]; }
+          { int32_t ij = i*(int32_t)n + j; M[ij] = A[ij] + v[j]; }
       }
   }
   
 void rmxn_shift_cols(uint32_t m, uint32_t n, double v[], double A[], double M[])
   { for (int32_t i = 0; i < m; i++)
       { for (int32_t j = 0; j < n; j++) 
-          { uint32_t ij = i*n + j; M[ij] = A[ij] + v[i]; }
+          { int32_t ij = i*(int32_t)n + j; M[ij] = A[ij] + v[i]; }
       }
   }
